+++ tools/sigclient/cmd/sigclient.go
@@ -104,32 +104,28 @@ var functionalTestCmd = &cobra.Command{
 	Use:   "functional",
 	Short: "functional testing of SigLens",
 	Run: func(cmd *cobra.Command, args []string) {
-		processCount, _ := cmd.Flags().GetInt("processCount")
 		dest, _ := cmd.Flags().GetString("dest")
 		queryDest, _ := cmd.Flags().GetString("queryDest")
-		totalEvents, _ := cmd.Flags().GetInt("totalEvents")
-		batchSize, _ := cmd.Flags().GetInt("batchSize")
-		indexPrefix, _ := cmd.Flags().GetString("indexPrefix")
-		numIndices, _ := cmd.Flags().GetInt("numIndices")
-		indexName, _ := cmd.Flags().GetString("indexName")
 		bearerToken, _ := cmd.Flags().GetString("bearerToken")
 		folderPath, _ := cmd.Flags().GetString("queryDataPath")
 
-		log.Infof("processCount : %+v\n", processCount)
+		totalEvents := 100_000
+		batchSize := 100
+		numIndices := 10
+		processCount := 1
+		indexPrefix := "ind"
+		indexName := ""
+		sleepDuration := 1 * time.Minute
+
 		log.Infof("dest : %+v\n", dest)
 		log.Infof("queryDest : %+v\n", queryDest)
-		log.Infof("totalEvents : %+v\n", totalEvents)
-		log.Infof("batchSize : %+v\n", batchSize)
-		log.Infof("indexPrefix : %+v\n", indexPrefix)
-		log.Infof("indexName : %+v\n", indexName)
-		log.Infof("numIndices : %+v\n", numIndices)
 		log.Infof("bearerToken : %+v\n", bearerToken)
 		log.Infof("queryDataPath : %+v\n", folderPath)
 
 		ingest.StartIngestion(ingest.ESBulk, "functional", "", totalEvents, false, batchSize, dest, indexPrefix,
 			indexName, numIndices, processCount, true, 0, bearerToken, 0, 0, nil)
 
-		time.Sleep(1 * time.Minute)
+		time.Sleep(sleepDuration)
 
 		query.FunctionalTest(queryDest, folderPath)
 	},
@@ -396,15 +392,15 @@ var traceCmd = &cobra.Command{
 
 func init() {
 	rootCmd.PersistentFlags().StringP("dest", "d", "", "Server URL.")
-	rootCmd.PersistentFlags().StringP("indexPrefix", "i", "ind", "index prefix")
-	rootCmd.PersistentFlags().StringP("indexName", "a", "", "index name")
 	rootCmd.PersistentFlags().StringP("bearerToken", "r", "", "Bearer token")
 
 	ingestCmd.PersistentFlags().IntP("processCount", "p", 1, "Number of parallel process to ingest data from.")
 	ingestCmd.PersistentFlags().IntP("totalEvents", "t", 1000000, "Total number of events to send")
 	ingestCmd.PersistentFlags().BoolP("continuous", "c", false, "Continous ingestion will ingore -t and will constantly send events as fast as possible")
 	ingestCmd.PersistentFlags().IntP("batchSize", "b", 100, "Batch size")
+	ingestCmd.PersistentFlags().StringP("indexPrefix", "i", "ind", "index prefix")
 	ingestCmd.PersistentFlags().Uint64P("eventsPerDay", "e", 0, "Number of events per day")
+	ingestCmd.PersistentFlags().StringP("indexName", "a", "", "index name")
 
 	esBulkCmd.Flags().BoolP("timestamp", "s", false, "Add timestamp in payload")
 	esBulkCmd.PersistentFlags().IntP("numIndices", "n", 1, "number of indices to ingest to")
@@ -414,10 +410,6 @@ func init() {
 	esBulkCmd.PersistentFlags().Uint32P("minColumns", "", 0, "minimum number of columns to generate. Default is 0. if 0, it will be set to maxColumns")
 	esBulkCmd.PersistentFlags().BoolP("enableVariableNumColumns", "", false, "generate a variable number of columns per record. Each record will have a random number of columns between minColumns and maxColumns")
 
-	functionalTestCmd.PersistentFlags().IntP("processCount", "p", 1, "Number of parallel process to ingest data from.")
-	functionalTestCmd.PersistentFlags().IntP("totalEvents", "t", 100000, "Total number of events to send")
-	functionalTestCmd.PersistentFlags().IntP("batchSize", "b", 100, "Batch size")
-	functionalTestCmd.PersistentFlags().IntP("numIndices", "n", 1, "number of indices to ingest to")
 	functionalTestCmd.PersistentFlags().StringP("queryDest", "q", "", "Query Server Address, format is IP:PORT")
 	functionalTestCmd.PersistentFlags().StringP("queryDataPath", "f", "", "path to folder containing query JSON files")
 
@@ -435,6 +427,7 @@ func init() {
 	queryCmd.PersistentFlags().StringP("outputFile", "o", "", "The filePath to output the Response time of Queries in csv format, When runResponseTime is set to True.")
 	queryCmd.PersistentFlags().BoolP("runResponseTime", "t", false, "Runs the given Queries and outputs the response time into a file in CSV format.")
 	queryCmd.PersistentFlags().StringSliceP("lookups", "l", []string{}, "List of lookups to migrate")
+	queryCmd.PersistentFlags().StringP("indexPrefix", "i", "ind", "index prefix")
 
 	traceCmd.PersistentFlags().StringP("filePrefix", "f", "", "Name of file to output to")
 	traceCmd.PersistentFlags().IntP("totalEvents", "t", 1000000, "Total number of traces to generate")
