+++ tools/sigclient/functionalQueries/groupby.json
