+++ tools/sigclient/functionalQueries/stats.json
